Add configuration for specific OSX Keychain item
Previously the OSX Keychain support looked for an Internet password
for the given host, account, and protocol. However, this means a
separate program for IMAP synching needs to have its own Keychain item
for its protocol, even though the password for the mail server will be
the same.

This change allows setting a specific generic password item in the
Keychain as the source, which allows multiple programs to use the same
item, reducing the amount of manual work for a password change.